requirements.txt
flask>=2.0.0
torch>=1.10.0
transformers>=4.15.0
python-dotenv>=0.19.0
flask-limiter>=2.4.0